requirements.txt
tapi
uvicorn
torch
transformers
pandas
peft
huggingface_hub
tqdm
pytesseract
pdf2image
sentence-transformers
faiss-cpu
